Fix Windows build warnings for VC 2022 (#8)

+++ code/point.cpp
@@ -68,7 +68,7 @@ float Point::DistanceTo(const Point &other) const {
   a = a * a;
   float term = sinf(deltaLng / 2);
   a += term * term * cosf(lat1) * cosf(lat2);
-  float c = 2 * atan2(sqrtf(a), sqrtf(1-a));
+  float c = 2 * atan2f(sqrtf(a), sqrtf(1-a));
   
   const float earthRadius = 6371000;
   return c * earthRadius;
